Break dependency loop in UHV headers
