+++ source/extensions/http/header_validators/envoy_default/BUILD
@@ -13,19 +13,66 @@ envoy_cc_library(
     name = "header_validator_common",
     srcs = [
         "header_validator.cc",
+    ],
+    hdrs = [
+        "header_validator.h",
+    ],
+    visibility = [
+        "//test/common/http/http1:__subpackages__",
+        "//test/extensions/http/header_validators/envoy_default:__subpackages__",
+    ],
+    deps = [
+        ":character_tables",
+        ":error_codes",
+        ":path_normalizer",
+        "//envoy/http:header_validator_interface",
+        "//external:abseil_node_hash_set",
+        "//source/common/http:headers_lib",
+        "@envoy_api//envoy/extensions/http/header_validators/envoy_default/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_library(
+    name = "character_tables",
+    hdrs = [
+        "character_tables.h",
+    ],
+    visibility = [
+        "//test/common/http/http1:__subpackages__",
+        "//test/extensions/http/header_validators/envoy_default:__subpackages__",
+    ],
+)
+
+envoy_cc_library(
+    name = "error_codes",
+    hdrs = [
+        "error_codes.h",
+    ],
+    visibility = [
+        "//test/common/http/http1:__subpackages__",
+        "//test/extensions/http/header_validators/envoy_default:__subpackages__",
+    ],
+    deps = [
+        "//source/common/singleton:const_singleton",
+    ],
+)
+
+envoy_cc_library(
+    name = "path_normalizer",
+    srcs = [
         "path_normalizer.cc",
     ],
     hdrs = [
-        "character_tables.h",
-        "header_validator.h",
         "path_normalizer.h",
     ],
     visibility = [
         "//test/common/http/http1:__subpackages__",
+        "//test/extensions/http/header_validators/envoy_default:__subpackages__",
     ],
     deps = [
+        ":character_tables",
+        ":error_codes",
         "//envoy/http:header_validator_interface",
-        "//external:abseil_node_hash_set",
         "//source/common/http:headers_lib",
         "@envoy_api//envoy/extensions/http/header_validators/envoy_default/v3:pkg_cc_proto",
     ],
@@ -41,8 +88,10 @@ envoy_cc_library(
     ],
     visibility = [
         "//test/common/http/http1:__subpackages__",
+        "//test/extensions/http/header_validators/envoy_default:__subpackages__",
     ],
     deps = [
+        ":error_codes",
         ":header_validator_common",
         "//envoy/http:header_validator_interface",
         "//external:abseil_node_hash_set",
@@ -61,8 +110,10 @@ envoy_cc_library(
     ],
     visibility = [
         "//test/common/http/http1:__subpackages__",
+        "//test/extensions/http/header_validators/envoy_default:__subpackages__",
     ],
     deps = [
+        ":error_codes",
         ":header_validator_common",
         "//envoy/http:header_validator_interface",
         "//external:abseil_node_hash_map",
