+++ source/extensions/http/header_validators/envoy_default/http1_header_validator.cc
@@ -2,6 +2,7 @@
 
 #include "source/common/http/utility.h"
 #include "source/extensions/http/header_validators/envoy_default/character_tables.h"
+#include "source/extensions/http/header_validators/envoy_default/error_codes.h"
 
 #include "absl/container/node_hash_set.h"
 #include "absl/strings/string_view.h"
