+++ test/extensions/http/header_validators/envoy_default/header_validator_test.cc
@@ -1,6 +1,7 @@
 #include "test/extensions/http/header_validators/envoy_default/header_validator_test.h"
 
 #include "source/extensions/http/header_validators/envoy_default/character_tables.h"
+#include "source/extensions/http/header_validators/envoy_default/error_codes.h"
 #include "source/extensions/http/header_validators/envoy_default/header_validator.h"
 
 namespace Envoy {
