+++ test/extensions/http/header_validators/envoy_default/http1_header_validator_test.cc
@@ -1,3 +1,4 @@
+#include "source/extensions/http/header_validators/envoy_default/error_codes.h"
 #include "source/extensions/http/header_validators/envoy_default/http1_header_validator.h"
 
 #include "test/extensions/http/header_validators/envoy_default/header_validator_test.h"
